Add save feature for Current list
Note that now Previous list requires new prop `firstMonday` that
eventually the user should provide on the first use of the extension.

Another thing to note is that the way `currentMonday` is implemented is
like a global variable that is created in `popup.js` entry point; so if
user clicks save when it's not currently the correct week, if they close
and open the extension again the Current tab will show the actual
current week. However because their data is stored in browser storage
they can simply click save again to see their new week's content.

Closes #12.